# docs/swagger.yaml
basePath: /
definitions:
...
  entities.Tag:
    type: object
...
  github_com_pksingh21_pyqbackend_entity.Question:
    type: object
  github_com_pksingh21_pyqbackend_entity.Test:
    type: object
host: localhost:9999
info:
  contact:
    email: fiber@swagger.io
    name: API Support
  description: This is a sample swagger for Fiber
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Fiber Example API
  version: "1.0"
paths:
...
  /questions:
    get:
      description: Retrieve all questions
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_pksingh21_pyqbackend_entity.Question'
            type: array
      summary: Get all questions
      tags:
      - questions
    post:
      consumes:
      - application/json
      description: Create a new question with provided details
      parameters:
      - description: Question data
        in: body
        name: question
        required: true
        schema:
          $ref: '#/definitions/github_com_pksingh21_pyqbackend_entity.Question'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_pksingh21_pyqbackend_entity.Question'
        "400":
          description: Bad Request
      summary: Create a new question
      tags:
      - questions
  /questions/{id}:
    delete:
      description: Delete a question by its ID
      parameters:
      - description: Question ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "404":
          description: Not Found
      summary: Delete a question by ID
      tags:
      - questions
    get:
      description: Retrieve a question by its ID
      parameters:
      - description: Question ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_pksingh21_pyqbackend_entity.Question'
        "404":
          description: Not Found
      summary: Get a question by ID
      tags:
      - questions
    put:
      consumes:
      - application/json
      description: Update an existing question by its ID
      parameters:
      - description: Question ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated question data
        in: body
        name: question
        required: true
        schema:
          $ref: '#/definitions/github_com_pksingh21_pyqbackend_entity.Question'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_pksingh21_pyqbackend_entity.Question'
        "400":
          description: Bad Request
        "404":
          description: Not Found
      summary: Update a question by ID
      tags:
      - questions
  /tags:
    get:
      description: Retrieve all tags
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entities.Tag'
            type: array
      summary: Get all tags
      tags:
      - tags
    post:
      consumes:
      - application/json
      description: Create a new tag with provided details
      parameters:
      - description: Tag data
        in: body
        name: tag
        required: true
        schema:
          $ref: '#/definitions/entities.Tag'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entities.Tag'
        "400":
          description: Bad Request
      summary: Create a new tag
      tags:
      - tags
  /tags/{id}:
    delete:
      description: Delete a tag by its ID
      parameters:
      - description: Tag ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "404":
          description: Not Found
      summary: Delete a tag by ID
      tags:
      - tags
    get:
      description: Retrieve a tag by its ID
      parameters:
      - description: Tag ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.Tag'
        "404":
          description: Not Found
      summary: Get a tag by ID
      tags:
      - tags
    put:
      consumes:
      - application/json
      description: Update an existing tag by its ID
      parameters:
      - description: Tag ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated tag data
        in: body
        name: tag
        required: true
        schema:
          $ref: '#/definitions/entities.Tag'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.Tag'
        "400":
          description: Bad Request
        "404":
          description: Not Found
      summary: Update a tag by ID
      tags:
      - tags
  /tests:
    get:
      description: Retrieve all tests
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_pksingh21_pyqbackend_entity.Test'
            type: array
      summary: Get all tests
      tags:
      - tests
    post:
      consumes:
      - application/json
      description: Create a new test with provided details
      parameters:
      - description: Test data
        in: body
        name: test
        required: true
        schema:
          $ref: '#/definitions/github_com_pksingh21_pyqbackend_entity.Test'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_pksingh21_pyqbackend_entity.Test'
        "400":
          description: Bad Request
      summary: Create a new test
      tags:
      - tests
  /tests/{id}:
    delete:
      description: Delete a test by its ID
      parameters:
      - description: Test ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "404":
          description: Not Found
      summary: Delete a test by ID
      tags:
      - tests
    get:
      description: Retrieve a test by its ID
      parameters:
      - description: Test ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_pksingh21_pyqbackend_entity.Test'
        "404":
          description: Not Found
      summary: Get a test by ID
      tags:
      - tests
    put:
      consumes:
      - application/json
      description: Update an existing test by its ID
      parameters:
      - description: Test ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated test data
        in: body
        name: test
        required: true
        schema:
          $ref: '#/definitions/github_com_pksingh21_pyqbackend_entity.Test'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_pksingh21_pyqbackend_entity.Test'
        "400":
          description: Bad Request
        "404":
          description: Not Found
      summary: Update a test by ID
      tags:
      - tests
...
swagger: "2.0"